server/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
Pillow==5.4.1
channels-redis==2.4.0
nose==1.3.7
pytest-asyncio==0.10.0
pytest-django==3.5.1
daphne==2.3.0